Add faster image processor
