+++ funcs/post_processing/images/soot_foil/deltas.py
@@ -1,5 +1,6 @@
+import os
+
 import numpy as np
-import uncertainties as un
 from skimage import io, color
 from uncertainties import unumpy as unp
 
@@ -9,10 +10,52 @@ from ._rust import fast_get_px_deltas_from_lines as _fast_get_deltas
 u_cell = u_cell["soot_foil"]
 
 
+class Shot:
+    def __init__(
+        self,
+        date,
+        shot_no,
+    ):
+        """
+        A cleaner-than-a-tuple organizer for date/shot info
+
+        Parameters
+        ----------
+        date: str
+            Shot date
+        shot_no: int
+            Shot number
+        """
+        self.date = date
+        self.shot_no = shot_no
+
+    def __repr__(self):
+        return f"{self.date} {self.shot_string}"
+
+    @property
+    def shot_string(self):
+        return f"Shot {self.shot_no:02d}"
+
+    @shot_string.setter
+    def shot_string(self, _):
+        raise AttributeError("shot_string cannot be set")
+
+    @property
+    def dir_name(self):
+        return os.path.join(self.date, self.shot_string)
+
+    @dir_name.setter
+    def dir_name(self, _):
+        raise AttributeError("dir_name cannot be set")
+
+    def __len__(self):
+        return len(str(self))
+
+
 def get_px_deltas_from_lines(
         img_path,
-        use_fast=True,
         mask_path=None,
+        use_fast=True,
         apply_uncertainty=True
 ):
     """
@@ -25,11 +68,10 @@ def get_px_deltas_from_lines(
     ----------
     img_path : str
         path of the image containing traced triple point lines
-    use_fast : bool
-        If true, the faster compiled rust version of the processing script will be run
     mask_path : Optional[str]
-        Path to the damage mask image, which must have the same dimensions as the main image. Only works if
-        `use_fast=True`
+        Path to the damage mask image, which must have the same dimensions as the main image
+    use_fast : bool
+        If true, the faster compiled rust version of the processing script will be run. Probably linux x86_64 only.
     apply_uncertainty : bool
         True returns array of nominal values with uncertainties; False returns
         only nominal values
@@ -43,12 +85,11 @@ def get_px_deltas_from_lines(
         deltas = _fast_get_deltas(img_path, mask_path)
     else:
         img = color.rgb2gray(io.imread(img_path))
-        img_max = img.max()
-        deltas = []
-        for i in range(img.shape[0]):
-            deltas.extend(_get_measurement_from_row(img[i], img_max))
-
-        deltas = np.array(deltas)
+        if mask_path:
+            mask = color.rgb2gray(io.imread(mask_path))
+        else:
+            mask = np.zeros_like(img)
+        deltas = _slow_get_deltas(img, mask)
 
     if apply_uncertainty:
         uncert = add_uncertainty_terms([
@@ -63,59 +104,145 @@ def get_px_deltas_from_lines(
     return deltas
 
 
-def _get_measurement_from_row(row, img_max):
-    row_locs = np.where(row == img_max)[0]
-    # get rid of adjacent pixels
-    # if two measurements are in adjacent pixels, this method selects the
-    # rightmost adjacent location and throws out the others (i.e. it only
-    # accepts measurements where the location is >1 away from the previous)
-    row_locs = row_locs[
-        np.abs(
-            np.diff(
-                [row_locs, np.roll(row_locs, -1)],
-                axis=0
-            )
-        ).flatten() > 1
-    ]
-    return np.diff(row_locs)
-
-
-def get_cell_size_from_deltas(
-        deltas,
-        l_px_i,
-        l_mm_i,
-        estimator=np.median
-):
+def _load_image(img_path):
     """
-    Converts pixel triple point deltas to cell size
+    Loads in an image as a black and white array
 
     Parameters
     ----------
-    deltas : np.array or pandas.Series
-    l_px_i : float
-        nominal value of spatial calibration factor (px)
-    l_mm_i : float
-        nominal value of spatial calibration factor (mm)
-    estimator : function
-        function used to estimate cell size from triple point measurements
+    img_path: str
+        Path of image to be loaded in.
 
     Returns
     -------
-    un.ufloat
-        estimated cell size
+
     """
-    l_px_i = un.ufloat(
-        l_px_i,
-        add_uncertainty_terms([
-            u_cell["l_px"]["b"],
-            u_cell["l_px"]["p"]
-        ])
+    return _binarize_array(color.rgb2gray(io.imread(img_path)))
+
+
+def _binarize_array(arr_in):
+    """
+    Converts an input array to have integer values of 1 at max, and 0
+    everywhere else.
+
+    Parameters
+    ----------
+    arr_in: np.ndarray
+        Input array
+
+    Returns
+    -------
+    np.ndarray
+        Converted array (copy)
+    """
+    arr_max = np.nanmax(arr_in)
+    # this can still be NaN if the input array is all NaN
+    if np.isnan(arr_max) or arr_max <= 0:
+        raise ValueError("cannot process image with max <= 0 or max == NaN")
+
+    arr_out = (arr_in.copy() / arr_in.max()).astype(int)
+    arr_out[arr_out < 1] = 0
+
+    return arr_out
+
+
+def _slow_get_deltas(
+    img,
+    mask,
+):
+    """
+    Returns an array of per-row triple point deltas (in pixels) from a given
+    image. The spatial calibration factor, mm_per_px, is required in order to
+    remove deltas larger than the tube diameter, which are unphysical and a
+    result of the limitations of soot foil measurement.
+
+    Parameters
+    ----------
+    img : np.ndarray
+        Array containing traced triple point lines for a single direction.
+    mask : np.ndarray
+        Array containing traced triple point lines for a single direction.
+
+    Returns
+    -------
+    deltas : np.ndarray
+        Triple point distances, in pixels.
+    """
+    _img = img.copy()
+
+    if mask.shape != _img.shape:
+        raise ValueError(f"image shape mismatch: {mask.shape} vs. {_img.shape}")
+    else:
+        _mask = mask.copy()
+
+    # NaN out exclusion zones
+    _img = np.where(_mask == 1, np.NaN, _img)
+
+    deltas = []
+    for i in range(_img.shape[0]):
+        diffs = _get_diffs_from_row(_img[i])
+        deltas.extend(diffs)
+
+    deltas = np.array(deltas)
+
+    return deltas
+
+
+def _get_diffs_from_sub_row(sub_row):
+    """
+    The actual diff-getter. If two measurements are in adjacent pixels, this
+    method selects the leftmost and rightmost adjacent locations and throws out
+    the others (i.e. it only accepts measurements where the boundary location
+    is >1 px away from the previous boundary location). The distance between
+    the leftmost and rightmost adjacent locations is not counted.
+
+    Parameters
+    ----------
+    sub_row: np.ndarray
+        Portion of a single row of image pixels within a given exclusion zone.
+
+    Returns
+    -------
+    np.ndarray
+        Distances between cell boundary locations.
+    """
+    # locate cell boundaries
+    cell_boundary_indices = np.where(sub_row == 1)[0]
+
+    # find how far apart adjacent boundaries are
+    cell_boundary_index_diffs = np.abs(
+        cell_boundary_indices - np.roll(cell_boundary_indices, -1)
+    )[:-1]  # throw out last diff -- roll wraps the first location around!
+
+    # throw out adjacent boundaries
+    cell_boundary_index_diffs = (
+        cell_boundary_index_diffs[cell_boundary_index_diffs > 1]
     )
-    l_mm_i = un.ufloat(
-        l_mm_i,
-        add_uncertainty_terms([
-            u_cell["l_mm"]["b"],
-            u_cell["l_mm"]["p"]
-        ])
-    )
-    return 2 * estimator(deltas) * l_mm_i / l_px_i
+
+    return cell_boundary_index_diffs
+
+
+def _get_diffs_from_row(row):
+    """
+    Get all pixel distances between cell boundaries for a single row in an image
+
+    Parameters
+    ----------
+    row: np.ndarray
+        Row containing 1 where cell boundaries exist, NaN where exclusion zones
+        exist, and 0 otherwise.
+
+    Returns
+    -------
+    np.ndarray
+        Concatenated array of cell boundary distances within exclusion zones.
+    """
+    diffs = []
+    # skip rows without useful data
+    if not (np.all(np.isnan(row)) or np.allclose(row, 0)):
+        # split into sub-arrays on NaN to enforce exclusion zones
+        split_row = np.split(row, np.where(np.isnan(row))[0])
+        for sub_row in split_row:
+            diffs.extend(_get_diffs_from_sub_row(sub_row))
+
+    return np.array(diffs)
