--- scripts/simulation_measurement_comparison/utilities/foil_image/exceptions.py
@@ -1,7 +1,0 @@
-"""
-Exceptions to be used within this module
-"""
-
-
-class ImageProcessingError(Exception):
-    pass
